Move some sharable setup from specs for WorkDerivativeLoader to shared context, for use in testing indexing, and other things.
